csv-ingestor/requirement.txt
requests
mysql-connector-python
tqdm
